+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/).
 
 ## [Unreleased]
 
+## [v0.7.2] - 2026-08-30
+
 ### Fixed
 
 - sleep timer now restores original volume when adding/setting time during fade-out
@@ -593,7 +595,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/).
 - Custom fonts support
 - Dynamic theme support
 
-[Unreleased]: https://github.com/likhithpraveenk/storii/compare/v0.7.1...HEAD
+[Unreleased]: https://github.com/likhithpraveenk/storii/compare/v0.7.2...HEAD
+[v0.7.2]: https://github.com/likhithpraveenk/storii/compare/v0.7.1...v0.7.2
 [v0.7.1]: https://github.com/likhithpraveenk/storii/compare/v0.7.0...v0.7.1
 [v0.7.0]: https://github.com/likhithpraveenk/storii/compare/v0.6.3...v0.7.0
 [v0.6.3]: https://github.com/likhithpraveenk/storii/compare/v0.6.2...v0.6.3